+++ CHANGELOG.rst
@@ -6,8 +6,13 @@ This document describes changes between each past release.
 3.4.0 (unreleased)
 ------------------
 
-- Rewrite e2e to use ``.get_records_timestamp()`` (#258)
+**Bug fixes**
 
+- Prevent events to be sent if status is not changed (#268)
+
+**Internal Changes**
+
+- Rewrite e2e to use ``.get_records_timestamp()`` (#258)
 
 3.3.3 (2018-10-10)
 ------------------