+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ This document describes changes between each past release.
 **Internal Changes**
 
 - Rewrite e2e to use ``.get_records_timestamp()`` (#258)
+- Enable Wheel distribution (fixes #271)
+
 
 3.3.3 (2018-10-10)
 ------------------